Changing AutoTrail to use state machines instead of Directed Acyclic Graphs for workflows.
Using a DAG to represent a workflow had a few limitations including:
1. Expressing different transition rules for different steps. Rules applied to all steps. Special rules for specific types of steps was difficult to implement.
2. End state determination was difficult. End state is the state where the workflow can no longer be run. This can happen if the various steps end up in a state where there is no way to move them forward in the DAG.

With these changes, each step is a state machine with its own states, actions and transition rules. A workflow is achieved by guarding actions on state machines with pre-conditions, which state the states *other* machines need to be in. An action can be performed when the other machines are in the states as specified in the pre-conditions. With this, one machine can be run only if its predecessors have successfully run, thereby achieving a workflow. However, since such pre-conditions can be specified for *any* state transition, arbitrary steps can be run based on transitions (e.g., if a step fails, a clean-up step can be run). Such flexibility with a DAG would be much more complicated.

End state determination is always deterministic. The end state is when all the machines have no possible actions, thereby, no state transitions are available. This simplifies workflow monitoring and management.

Defining APIs is greatly simplified by simpler API handler requirements. There is first-class support for exposing the context via API. This means, just like operations can be performed on the machines via API, so is it for the context object.

There are no unit tests as they are too tightly integrated with the implementation and often require changes along with any code change. However, there are integration/validation tests that run workflows, perform API calls and verify functionality achieving over 80% code coverage currently.